teporarily clear screen at each frame.
This is inefficient and slow, but will make creating the game mechanics easier. Not least because it's so slow...
In future, go back to only draw the parts of the screen that has
changed. However, care must be taken to clear the previous position
since last draw. Otherwise the object will expand.